+++ src/PartnerCenter/Invoices/InvoiceOperations.cs
@@ -55,7 +55,7 @@ namespace Microsoft.Store.PartnerCenter.Invoices
         /// <param name="period">The period for unbilled recon.</param>
         /// <param name="size">The page size.</param>
         /// <returns>The recon line item collection operations.</returns>
-        public IReconLineItemCollection By(BillingProvider provider, InvoiceLineItemType invoiceLineItemType, string currencyCode, UnbilledPeriod period, int? size)
+        public IReconLineItemCollection By(BillingProvider provider, InvoiceLineItemType invoiceLineItemType, string currencyCode, BillingPeriod period, int? size)
             => new ReconLineItemCollectionOperations(Partner, Context, provider, invoiceLineItemType, currencyCode, period, size);
      
         /// <summary>